fixed a button bug
had a bug were the button for placing the tower disapeared so had to fix it xD
just changed the jpg to a png and reimported it and correctly placed it within the main scene.

also fixed a connecting node issue since the new style of doing things is slightly diffrent with the latest godot update.